feat: test infrastructure — pool 1000 utenti, simulazione, analysis panel
- SQL: create_test_pool(), simulate_airdrop_participation(), get_airdrop_analysis()
- SQL: fix ROUND cast NUMERIC, get_draw_preview con p_service_call bypass
- SQL: leaderboard esclude test users, create_test_pool upsert conflict
- FE: admin panel Test Pool + Airdrop Analysis con dettaglio score/NFT/split
- FE: simulazione blocchi in modal airdrop (slider %, spread giorni)
- FE: stats admin escludono is_test_user
- report.js: esclude test users da conteggi

+++ supabase/migrations/20260319221008_airdrop_engine_rpc.sql
@@ -152,10 +152,10 @@ BEGIN
       f1c.user_id,
       f1c.total_blocks,
       f1c.total_aria,
-      ROUND(f1c.f1, 6) AS f1,
-      ROUND(f2c.f2, 6) AS f2,
-      ROUND(f3c.f3, 6) AS f3,
-      ROUND(v_w1 * f1c.f1 + v_w2 * f2c.f2 + v_w3 * f3c.f3, 6) AS score
+      ROUND(f1c.f1::NUMERIC, 6) AS f1,
+      ROUND(f2c.f2::NUMERIC, 6) AS f2,
+      ROUND(f3c.f3::NUMERIC, 6) AS f3,
+      ROUND((v_w1 * f1c.f1 + v_w2 * f2c.f2 + v_w3 * f3c.f3)::NUMERIC, 6) AS score
     FROM f1_calc f1c
     JOIN f2_calc f2c ON f2c.user_id = f1c.user_id
     JOIN f3_calc f3c ON f3c.user_id = f1c.user_id
